Run the icinga config wizard on osticket, saved data in docs.
Here I'm checking if there's a feature that I'd expect to be here. I'd
assume that if there is a longer commit description after an empty
line, then the further description would be the long version of the
git commit message. Fingers crossed! need to check how it looks on
github.