Adding more checks for osticket. Trying to fix errors.
As I've added new checks for the osticket system, there were some
errors. First thought it was a typo in hosts.conf, but turned out that
the osticket zone had an example of osticke instead of osticket. After
fixing that there was another problem with parent zone. Apparently the
way zones are created changed since version 2.5. Tried changing the
parent zone, but it resulted with another error. I'll continue the
work tomorrow.